requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
#scikit-learn==1.3.0
redis==4.5.5
python-dotenv==1.0.0
schedule==1.2.0
joblib==1.3.2
requests==2.31.0
supabase==2.3.1
numpy==1.23.5
scikit-learn==1.2.2